--- a/.gitignore
+++ b/.gitignore
@@ -10,3 +10,4 @@ dist/
 results/
 
 .coverage
+.vscode