### Description

This PR adds a reference implementation, kernel, op, test, and benchmark for Non-Max Suppression (NMS).
The implementation of this kernel is [based on the CUDA reference implementation from Torchvision](https://github.com/pytorch/vision/blob/0721867e42841171254c7acaa45fbaf8ee16d3d7/torchvision/csrc/ops/cuda/nms_kernel.cu).
Our operation also follows [the same interface as the Torchvision API](https://github.com/pytorch/vision/blob/0721867e42841171254c7acaa45fbaf8ee16d3d7/torchvision/ops/boxes.py#L20).

We took a slightly different approach when implementing this kernel; first, we wrote reference PyTorch implementations and tried to use `torch.compile()` to accelerate them. The performance of the `compile()`-d kernels was very poor compared to the CUDA reference (this is somewhat unsurprising, given the parallelization structure of NMS). Next, we implemented the handwritten Triton kernel. Finally, we introduced as many optimizations as possible from the CUDA reference impl.

Our "final" Triton kernel is generally still a bit slower than CUDA, especially for large problem sizes. This is primarily because of two differences in the implementations:
1. The CUDA kernel uses a bitmask to represent the IoU for each column block, which saves the amount of scratch memory that the kernel requires (`N * num_col_blocks` for CUDA vs. `N * N` for Triton).
2. The CUDA kernel uses shared memory to represent the removed mask in the stage2 kernel, whereas the Triton kernel must use HBM. This makes the lookups for whether or not each box has been kept much slower in Triton, and forces us to synchronize after each loop iteration

### Testing

Please select all that apply.

- [ ] Existing unit tests
- [x] Unit tests added by this PR
- [ ] Other (please explain)
- [ ] This PR is not tested

#### Test instructions

```
# Can also add `CONCH_ENABLE_TORCHVISION=1`
pytest tests/nms_test.py
```

#### Platforms

Please select all hardware platforms that this PR was tested on.

- [x] Nvidia GPU
- [x] AMD GPU
- [ ] Other (please explain)

### Benchmarks

```
python benchmarks/nms_benchmark.py --num-boxes 4096 --gpu-ref --compile-ref --vectorize-ref
```

**A10**

```
Reference vs Conch: Results matched with atol=0.001 :)
Parameters: {'num_boxes': 4096, 'iou_threshold': 0.5}
Conch: num_iterations=2756, min=3.076 ms, max=3.310 ms, mean=3.096 ms, median=3.091 ms
Baseline: num_iterations=13, min=642.600 ms, max=755.242 ms, mean=716.487 ms, median=726.104 ms
PyTorch Reference (Compiled): num_iterations=23, min=404.111 ms, max=466.083 ms, mean=424.135 ms, median=420.500 ms
PyTorch Reference (Vectorized): num_iterations=1, min=165183.766 ms, max=165183.766 ms, mean=165183.766 ms, median=165183.766 ms
PyTorch Reference (Vectorized, Compiled): num_iterations=1, min=181160.094 ms, max=181160.094 ms, mean=181160.094 ms, median=181160.094 ms
PyTorch GPU Reference: num_iterations=5705, min=1.181 ms, max=1.473 ms, mean=1.213 ms, median=1.211 ms
```

**H100**

```
Reference vs Conch: Results matched with atol=0.001 :)
Parameters: {'num_boxes': 4096, 'iou_threshold': 0.5}
Conch: num_iterations=6224, min=1.479 ms, max=2.487 ms, mean=1.507 ms, median=1.503 ms
Baseline: num_iterations=21, min=457.674 ms, max=517.613 ms, mean=463.726 ms, median=460.260 ms
PyTorch Reference (Compiled): num_iterations=33, min=293.638 ms, max=359.811 ms, mean=301.033 ms, median=296.215 ms
PyTorch Reference (Vectorized): num_iterations=1, min=123716.055 ms, max=123716.055 ms, mean=123716.055 ms, median=123716.055 ms
PyTorch Reference (Vectorized, Compiled): num_iterations=1, min=124315.828 ms, max=124315.828 ms, mean=124315.828 ms, median=124315.828 ms
PyTorch GPU Reference: num_iterations=11380, min=0.743 ms, max=5.353 ms, mean=0.764 ms, median=0.761 ms
```

**MI300X**

```
Reference vs Conch: Results matched with atol=0.001 :)
Parameters: {'num_boxes': 4096, 'iou_threshold': 0.5}
Conch: num_iterations=5374, min=1.803 ms, max=11.276 ms, mean=1.829 ms, median=1.822 ms
Baseline: num_iterations=22, min=441.022 ms, max=455.029 ms, mean=445.137 ms, median=443.703 ms
PyTorch Reference (Compiled): num_iterations=37, min=260.512 ms, max=274.488 ms, mean=263.063 ms, median=262.745 ms
PyTorch Reference (Vectorized): num_iterations=1, min=137248.344 ms, max=137248.344 ms, mean=137248.344 ms, median=137248.344 ms
PyTorch Reference (Vectorized, Compiled): num_iterations=1, min=137764.391 ms, max=137764.391 ms, mean=137764.391 ms, median=137764.391 ms
PyTorch GPU Reference: num_iterations=10252, min=0.903 ms, max=1.466 ms, mean=0.921 ms, median=0.920 ms
```